pyproject.toml
[tool.poetry]
name = "snypr"
version = "0.1.0"
description = ""
authors = ["lazy-geeek <lazy-geek@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.8"
pandas = "^2.2.2"
requests = "^2.32.3"
telethon = "^1.35.0"
websockets = "^12.0"
fake-useragent = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
